extract filesystem test utils
